requirements.txt
fastapi[standard]==0.115.12
uvicorn[standard]==0.34.2
jinja2==3.1.6
aiosqlite==0.21.0
httpx==0.28.1
feedparser==6.0.11
apscheduler==3.11.0
pydantic-settings==2.9.1
google-genai==1.14.0
python-multipart==0.0.20
